feat: 구매한 로또 장수 계산
feat: 구매한 로또 장수 계산

1000원 단위 아니면 예외 처리하기

+++ src/main/java/lotto/Console.java
@@ -2,9 +2,9 @@ package lotto;
 import static camp.nextstep.edu.missionutils.Console.readLine;
 public class Console {
 
-    public double inputMoney() {
+    public int inputMoney() {
         System.out.println("구입금액을 입력해 주세요.");
-        double money = Double.valueOf(readLine());
+        int money = Integer.valueOf(readLine());
         return money;
     }
 }
